Refactored out the list accounts method
