+++ src/api.rs
@@ -8,21 +8,9 @@ use serde_json::json;
 use r2d2::Pool;
 use r2d2_sqlite::SqliteConnectionManager;
 
-use crate::account::data::AccountType;
+use crate::account::data::{Account, AccountType};
 use crate::datastruct;
 use crate::db;
-
-pub fn index(
-    pool: web::Data<Pool<SqliteConnectionManager>>,
-) -> impl Future<Item = HttpResponse, Error = Error> {
-    let conn = pool.get().unwrap();
-    let result = db::list_accounts(conn);
-
-    match result {
-        Ok(v) => ok(HttpResponse::Ok().json(v)),
-        Err(_e) => ok(HttpResponse::InternalServerError().finish()),
-    }
-}
 
 pub fn list_transactions() -> impl Future<Item = HttpResponse, Error = Error> {
     let result = db::list_transactions();
@@ -84,7 +72,7 @@ pub fn get_transactions_date_scoped(
     }
 }
 
-fn are_accounts_compatible(from: &datastruct::Account, to: &datastruct::Account) -> bool {
+fn are_accounts_compatible(from: &Account, to: &Account) -> bool {
     if &from.id == &to.id || &from.currency != &to.currency {
         return false;
     }
@@ -234,17 +222,6 @@ pub fn get_account_balance(
     }
 }
 
-pub fn list_expense_accounts(
-    pool: web::Data<Pool<SqliteConnectionManager>>,
-) -> impl Future<Item = HttpResponse, Error = Error> {
-    let result = db::list_accounts_filter_type(pool.get().unwrap(), AccountType::Expenses);
-
-    match result {
-        Ok(v) => ok(HttpResponse::Ok().json(v)),
-        Err(_e) => ok(HttpResponse::InternalServerError().finish()),
-    }
-}
-
 pub fn list_currencies(
     pool: web::Data<Pool<SqliteConnectionManager>>,
 ) -> impl Future<Item = HttpResponse, Error = Error> {
@@ -277,8 +254,8 @@ pub fn check_ledger_integrity(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::account::data::Account;
     use crate::account::data::AccountType;
-    use crate::datastruct::Account;
 
     #[test]
     fn accounts_compatible() {
